MoveIt partially working
The move_to_cup.py file will eventually be the node for grasping the
cup. Right now it has a bunch of code for figuring out how to use
Baxter's built in IK solver. The solution of that solver (a set of joint
states) will go into a moveit planner. So far, I've gotten the planner
to work for planning with a joint position target, but not a pose
target.

I edited baxter_moveit_config.launch so that the gui doesn't pop up.
Also, I just changed to positions of some objects in the
test_world.world.

transforms.py is a script provided by Jarvis for doing conversions
between rotation matrices, se(3), and quaternions. It will be useful.