move_to_cup has good IK solution
I finally found a seed that will provide a solution that works. I think
part of the problem was that I kept trying to get to poses that were
outof range. Important: the measurement for z in the base frame is not
from the ground, but from the position of the base frame on Baxter -
about the middle of the chest. So far a good z position value for
grasping things on the gazebo table is .1